+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Yii Framework 2 gii extension Change Log
 - Bug #327: Fixed bug in Model generator when $baseClass is an abstract class (rhertogh)
 - Enh #366: Option to allow standardized class names capitals in model generator (slinstj)
 - Enh #378: Remove useless import of `Yii` from CRUD generator search model template (CeBe)
+- Bug #383: Fix incorrect title generation in CRUD update view (bscheshirwork)
 
 
 2.0.7 May 3, 2018