+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Yii Framework 2 gii extension Change Log
 - Enh #378: Remove useless import of `Yii` from CRUD generator search model template (CeBe)
 - Bug #383: Fix incorrect title generation in CRUD update view (bscheshirwork)
 
+
 2.0.7 May 3, 2018
 -----------------